backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
motor>=3.3.2
pymongo>=4.6.1
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
yfinance>=0.2.35
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.0
python-dotenv>=1.0.0
httpx>=0.26.0
bcrypt>=4.0.0
